refactor: rename util.js to base.js
