+++ ui/jobs/components/base.js
